Sredjena Student forma i napravljenje funkcije za vracanje, dodavanje, brisanje i azuriranje studenata u DTOManager

+++ Aplikacija/SBP-faza2/SBP-faza2/DTOs.cs
@@ -4,14 +4,14 @@
 #region Student
 public class StudentBasic
 {
-    public required int Id { get; set; }
+    public required long Id { get; set; }
     public required string LicnoIme { get; set; }
     public required string ImeRoditelja { get; set; }
     public required string Prezime { get; set; }
     public required string BrojIndeksa { get; set; }
     public required string Smer { get; set; }
 
-    public StudentBasic(int id, string licnoIme, string imeRoditelja, string prezime, string brojIndeksa, string smer) 
+    public StudentBasic(long id, string licnoIme, string imeRoditelja, string prezime, string brojIndeksa, string smer) 
     {
         Id = id;
         LicnoIme = licnoIme;
@@ -32,13 +32,13 @@ public class StudentBasic
 #region Predmet
 public class PredmetBasic
 {
-    public required int Id { get; set; }
+    public required long Id { get; set; }
     public required string Sifra { get; set; }
     public required string Naziv { get; set; }
     public required string Katedra { get; set; }
     public required string Semestar { get; set; }
 
-    public PredmetBasic(int id, string sifra, string naziv, string katedra, string semestar)
+    public PredmetBasic(long id, string sifra, string naziv, string katedra, string semestar)
     {
         Id = id;
         Sifra = sifra;
